requirements.txt
beautifulsoup4==4.12.3
langchain==0.3.7
langchain_community==0.3.5
langchain_core==0.3.15
langchain_openai==0.2.6
langchain_text_splitters==0.3.2
matplotlib==3.8.2
pandas==2.2.3
Pillow==10.1.0
Pillow==11.0.0
python-dotenv==1.0.1
Requests==2.32.3
streamlit==1.31.1
